remove a loop over the nodes
